api/requirements.txt
einops
fastapi
huggingface_hub
insightface
numpy
onnxruntime
opencv-python-headless
pillow
python-multipart
requests
safetensors
scikit-image
torch
torchvision
uvicorn
sqlalchemy
psycopg2-binary
alembic
passlib
python-jose